+++ app/routers/post_routers.py
@@ -1,0 +1,8 @@
+from fastapi import APIRouter
+
+router = APIRouter(prefix='/api/posts')
+
+
+@router.post('/')
+def create_post():
+    return {"message": "Post created"}
