--- app/schemas/sw_notice.py
@@ -1,23 +1,0 @@
-from datetime import datetime
-
-class SWNotice:
-    def __init__(self, title: str, created_date: datetime, link: str):
-        self.__title = title
-        self.__created_date = created_date
-        self.__link = link
-
-    @property
-    def title(self) -> str:
-        return self.__title
-
-    @property
-    def created_date(self) -> datetime:
-        return self.__created_date
-
-    @property
-    def link(self) -> str:
-        return self.__link
-
-    def __repr__(self) -> str:
-        return (f"SWNotice(title={self.__title!r}, "
-                f"created_date={self.__created_date!r}, link={self.__link!r})")
